+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 4270efd79eaa48f24ed2d430e731cb7e0ff95c6f
+// Checksum: 76a5e462c0095a314ea445f2c236d7f046f7b7b1
 //
 // ignore_for_file: unused_import
 
@@ -1841,6 +1841,18 @@ final class _EvaluateVisitor
   Value? visitIncludeRule(IncludeRule node) {
     var mixin = _addExceptionSpan(node,
         () => _environment.getMixin(node.name, namespace: node.namespace));
+    if (node.originalName.startsWith('--') &&
+        mixin is UserDefinedCallable &&
+        !mixin.declaration.originalName.startsWith('--')) {
+      _warn(
+          'Sass @mixin names beginning with -- are deprecated for forward-'
+          'compatibility with plain CSS mixins.\n'
+          '\n'
+          'For details, see https://sass-lang.com/d/css-function-mixin',
+          node.nameSpan,
+          Deprecation.cssFunctionMixin);
+    }
+
     var contentCallable = node.content.andThen((content) => UserDefinedCallable(
         content, _environment.closure(),
         inDependency: _inDependency));
@@ -1971,6 +1983,9 @@ final class _EvaluateVisitor
     if (_declarationName != null) {
       throw _exception(
           "Style rules may not be used within nested declarations.", node.span);
+    } else if (_inKeyframes && _parent is CssKeyframeBlock) {
+      throw _exception(
+          "Style rules may not be used within keyframe blocks.", node.span);
     }
 
     var (selectorText, selectorMap) =
@@ -2477,6 +2492,19 @@ final class _EvaluateVisitor
 
       function = (_stylesheet.plainCss ? null : _builtInFunctions[node.name]) ??
           PlainCssCallable(node.originalName);
+    }
+
+    if (node.originalName.startsWith('--') &&
+        function is UserDefinedCallable &&
+        !function.declaration.originalName.startsWith('--')) {
+      _warn(
+        'Sass @function names beginning with -- are deprecated for forward-'
+        'compatibility with plain CSS functions.\n'
+        '\n'
+        'For details, see https://sass-lang.com/d/css-function-mixin',
+        node.nameSpan,
+        Deprecation.cssFunctionMixin,
+      );
     }
 
     var oldInFunction = _inFunction;
@@ -3360,6 +3388,9 @@ final class _EvaluateVisitor
     if (_declarationName != null) {
       throw _exception(
           "Style rules may not be used within nested declarations.", node.span);
+    } else if (_inKeyframes && _parent is CssKeyframeBlock) {
+      throw _exception(
+          "Style rules may not be used within keyframe blocks.", node.span);
     }
 
     var styleRule = _styleRule;