+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.76.0
+## 1.78.0
 
 * **Breaking change**: Passing a number with unit `%` to the `$alpha` parameter
   of `color.change()`, `color.adjust()`, `change-color()`, and `adjust-color()`
@@ -183,6 +183,44 @@
   defined in Color Level 4.
 
 * Remove `RgbColor`, `HslColor` and `HwbColor` SassScript values.
+
+## 1.77.3
+
+### Dart API
+
+* `Deprecation.duplicateVariableFlags` has been deprecated and replaced with
+  `Deprecation.duplicateVarFlags` to make it consistent with the
+  `duplicate-var-flags` name used on the command line and in the JS API.
+
+## 1.77.2
+
+* Don't emit deprecation warnings for functions and mixins beginning with `__`.
+
+* Allow user-defined functions whose names begin with `_` and otherwise look
+  like vendor-prefixed functions with special CSS syntax.
+
+### Command-Line Interface
+
+* Properly handle the `--silence-deprecation` flag.
+
+* Handle the `--fatal-deprecation` and `--future-deprecation` flags for
+  `--interactive` mode.
+
+## 1.77.1
+
+* Fix a crash that could come up with importers in certain contexts.
+
+## 1.77.0
+
+* *Don't* throw errors for at-rules in keyframe blocks.
+
+## 1.76.0
+
+* Throw errors for misplaced statements in keyframe blocks.
+
+* Mixins and functions whose names begin with `--` are now deprecated for
+  forwards-compatibility with the in-progress CSS functions and mixins spec.
+  This deprecation is named `css-function-mixin`.
 
 ## 1.75.0